Remove legacy modules and simplify project structure

+++ ruoyi-sensor/src/main/java/com/ruoyi/sensor/domain/dto/SensorIngestAssociateRequest.java
@@ -7,6 +7,6 @@ public class SensorIngestAssociateRequest
 {
     private Long deviceId;
     private Long pointId;
-    /** phm_acquisition_channel primary key, not the physical channel number. */
-    private Long channelId;
+    /** Physical MAT acquisition channel number. */
+    private Integer channelId;
 }
